添加elasticsearch服务 编写查询商品相关信息

+++ vinspier/vinspier-common/src/main/java/com/vinspier/common/pojo/PageResult.java
@@ -7,7 +7,7 @@ public class PageResult<T> {
     // 总条数
     private Long total;
     // 总页数
-    private Long totalPage;
+    private Integer totalPage;
     // 总页数
     private List<T> items;// 当前页数据
 
@@ -19,7 +19,7 @@ public class PageResult<T> {
         this.items = items;
     }
 
-    public PageResult(Long total, Long totalPage, List<T> items) {
+    public PageResult(Long total, Integer totalPage, List<T> items) {
         this.total = total;
         this.totalPage = totalPage;
         this.items = items;
@@ -41,11 +41,11 @@ public class PageResult<T> {
         this.items = items;
     }
 
-    public Long getTotalPage() {
+    public Integer getTotalPage() {
         return totalPage;
     }
 
-    public void setTotalPage(Long totalPage) {
+    public void setTotalPage(Integer totalPage) {
         this.totalPage = totalPage;
     }
